fixes: fixed some bugs
