+++ cmd/client/src/components/Banner.tsx
